refactor: extract withTimeout to shared utility, fix clearTimeout leak
Move withTimeout helper to utils/with-timeout.ts and reuse in
apply-numstat.ts. The previous Promise.race approach in apply-numstat
leaked setTimeout timers on successful completion.
