+++ apps/desktop/src/lib/trpc/routers/changes/utils/with-timeout.ts
@@ -1,0 +1,25 @@
+/**
+ * Race a promise against a timeout. Clears the timer on both resolve and reject
+ * to avoid leaked timers.
+ */
+export function withTimeout<T>(
+	promise: Promise<T>,
+	ms: number,
+	label: string,
+): Promise<T> {
+	return new Promise<T>((resolve, reject) => {
+		const timer = setTimeout(() => {
+			reject(new Error(`git operation timed out after ${ms}ms: ${label}`));
+		}, ms);
+		promise.then(
+			(value) => {
+				clearTimeout(timer);
+				resolve(value);
+			},
+			(error) => {
+				clearTimeout(timer);
+				reject(error);
+			},
+		);
+	});
+}
